научился видеть на сервере, какое действие хочет быть выполненным по кнопке

+++ src/main/java/com/example/demo/web/rest/EventController.java
@@ -20,9 +20,15 @@ import org.springframework.web.bind.annotation.PostMapping;
 import com.example.demo.entity.User;
 import com.example.demo.service.SecurityService;
 import com.example.demo.service.UserService;
+import org.springframework.web.servlet.HttpServletBean;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.swing.text.html.parser.Entity;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.*;
 
 @Controller
@@ -210,18 +216,31 @@ public class EventController {
         return "redirect:/notMyEventUnsubscribed";
     }
 
-    @GetMapping("/event{Id}/delete")
-    public String delete(@PathVariable String Id, Model model) {
-        String username = securityService.findLoggedInUsername();
-        User user = userRepository.findByUsername(username);
+//    @GetMapping("/event{Id}/delete")
+//    public String delete(@PathVariable String Id, Model model) {
+//        String username = securityService.findLoggedInUsername();
+//        User user = userRepository.findByUsername(username);
+//
+//        var longId = Long.parseLong(Id);
+//        Event event = eventRepository.findById(longId).get();
+//        user.getSubscribeTo().remove(event); // проработать после того, как решится вопрос со связностью таблиц, нужно ли еще добавлять юзера в список для ивента
+//        userRepository.save(user);
+//        eventRepository.delete(event);
+//
+//        return "menu";
+//    }
 
-        var longId = Long.parseLong(Id);
-        Event event = eventRepository.findById(longId).get();
-        user.getSubscribeTo().remove(event); // проработать после того, как решится вопрос со связностью таблиц, нужно ли еще добавлять юзера в список для ивента
-        userRepository.save(user);
-        eventRepository.delete(event);
-
-        return "menu";
+    @PostMapping("/event{Id}")
+    public String postProcessing(@PathVariable String Id, HttpServletRequest request) throws IOException {
+        String payload = getRequestBody(request);
+        String actionType = payload.substring(11, payload.length());
+        if (actionType.equals("delete")) {
+            var longId = Long.parseLong(Id);
+            Event event = eventRepository.findById(longId).get();
+            eventRepository.delete(event);
+            return "menu";
+        }
+        return "redirect:/event" + Id;
     }
 
     @GetMapping("/editEvent{Id}")
@@ -287,7 +306,39 @@ public class EventController {
 //    }
 //
 
+    private static String getRequestBody(HttpServletRequest request) throws IOException {
 
+        String body = null;
+        StringBuilder stringBuilder = new StringBuilder();
+        BufferedReader bufferedReader = null;
+
+        try {
+            InputStream inputStream = request.getInputStream();
+            if (inputStream != null) {
+                bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
+                char[] charBuffer = new char[128];
+                int bytesRead = -1;
+                while ((bytesRead = bufferedReader.read(charBuffer)) > 0) {
+                    stringBuilder.append(charBuffer, 0, bytesRead);
+                }
+            } else {
+                stringBuilder.append("");
+            }
+        } catch (IOException ex) {
+            throw ex;
+        } finally {
+            if (bufferedReader != null) {
+                try {
+                    bufferedReader.close();
+                } catch (IOException ex) {
+                    throw ex;
+                }
+            }
+        }
+
+        body = stringBuilder.toString();
+        return body;
+    }
 
 
 }
